フォルダやファイルがあるか確認するかん関数使った。

+++ remove_files.py
@@ -3,9 +3,7 @@ import os
 for y in range(x):
   print(y)
   path='./txt/'+str(y)+'.txt'
-  # path=str(y)+'.txt'
-  os.remove(path)
-  # f=open(path,'w')
-  # f.write('')#ファイル作成
-  # f.close()
-os.rmdir('./txt')
+  if(os.path.isfile(path)):
+    os.remove(path)
+if not os.listdir('./txt'):
+  os.rmdir('./txt')
